API 서버 정수형 int -> Int32, long -> Int64로 전환

+++ APIServer/Service/impl/AttendanceService.cs
@@ -9,7 +9,7 @@ public class AttendanceService : IAttendanceService
     }
 
     public record AttendanceResult(ErrorCodes errorCode, bool isSuccess);
-    public async Task<AttendanceResult> AttendanceCheck(int userId, DateTime nowTime)
+    public async Task<AttendanceResult> AttendanceCheck(Int64 userId, DateTime nowTime)
     {
         // 서버와의 시간을 체크
         try
